Arithmetic library's flag argument converted to int type
The `flags' arguments to `gal_arithmetic' had an `unsigned char'
type. However, on some LLVM compilers, we got the following warning when
using `va_start': "promotion to 'va_start' has undefined behavior" and
"note: parameter of type 'unsigned char' is declared here" (pointing to the
`flags' argument of `gal_arithmetic').

To avoid this warning and have a clean build, in the arithmetic library,
the flags now have an `int' type.

This warning was first reported by Raúl Infante Sainz.

+++ lib/arithmetic-binary.c
@@ -348,8 +348,7 @@ along with Gnuastro. If not, see <http://www.gnu.org/licenses/>.
 /*************              Top level function          *****************/
 /************************************************************************/
 gal_data_t *
-arithmetic_binary(int operator, uint8_t flags, gal_data_t *lo,
-                  gal_data_t *ro)
+arithmetic_binary(int operator, int flags, gal_data_t *lo, gal_data_t *ro)
 {
   /* Read the variable arguments. `lo' and `ro' keep the original data, in
      case their type isn't built (based on configure options are configure
